+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/MessageConverter.kt
@@ -6,6 +6,7 @@ import com.google.gson.JsonObject
 import jp.shiguredo.sora.sdk.channel.option.SoraChannelRole
 import jp.shiguredo.sora.sdk.channel.option.SoraForwardingFilterOption
 import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
+import jp.shiguredo.sora.sdk.channel.option.SoraVideoOption
 import jp.shiguredo.sora.sdk.error.SoraDisconnectReason
 import org.webrtc.RTCStats
 import org.webrtc.RTCStatsReport
@@ -78,14 +79,22 @@ class MessageConverter {
                 } else {
                     msg.audio = false
                 }
+
                 if (mediaOption.videoUpstreamEnabled) {
-                    val videoSetting = VideoSetting(mediaOption.videoCodec.toString())
-                    mediaOption.videoBitrate?.let { videoSetting.bitRate = it }
-                    mediaOption.videoVp9Params?.let { videoSetting.vp9Params = it }
-                    mediaOption.videoAv1Params?.let { videoSetting.av1Params = it }
-                    mediaOption.videoH264Params?.let { videoSetting.h264Params = it }
-                    msg.video = videoSetting
+                    // video 関連設定がすべてデフォルト値の場合は video フィールドの設定を省略する
+                    if (!mediaOption.isDefaultVideoOption()) {
+                        msg.video = VideoSetting().apply {
+                            if (mediaOption.videoCodec != SoraVideoOption.Codec.DEFAULT) {
+                                codecType = mediaOption.videoCodec.toString()
+                            }
+                            mediaOption.videoBitrate?.let { bitRate = it }
+                            mediaOption.videoVp9Params?.let { vp9Params = it }
+                            mediaOption.videoAv1Params?.let { av1Params = it }
+                            mediaOption.videoH264Params?.let { h264Params = it }
+                        }
+                    }
                 } else {
+                    // ビデオを無効化したいため false を設定する
                     msg.video = false
                 }
             } else {
@@ -98,11 +107,19 @@ class MessageConverter {
                 } else {
                     msg.audio = false
                 }
+
                 if (mediaOption.videoDownstreamEnabled) {
-                    val videoSetting = VideoSetting(mediaOption.videoCodec.toString())
-                    // TODO(shino): 視聴側の bit_rate 設定はサーバで無視される
-                    mediaOption.videoBitrate?.let { videoSetting.bitRate = it }
-                    msg.video = videoSetting
+                    // video 関連設定がすべてデフォルト値の場合は video フィールドの設定を省略する
+                    if (!mediaOption.isDefaultVideoOption()) {
+                        msg.video = VideoSetting().apply {
+                            if (mediaOption.videoCodec != SoraVideoOption.Codec.DEFAULT) {
+                                codecType = mediaOption.videoCodec.toString()
+                            }
+                            // TODO(shino): 視聴側の bit_rate 設定はサーバで無視される
+                            // TODO(zztkm): ビデオコーデック以外は配信者が設定できる項目で、視聴者は設定不要なので、設定不要な項目は省略する
+                            mediaOption.videoBitrate?.let { bitRate = it }
+                        }
+                    }
                 } else {
                     msg.video = false
                 }